+++ README.md
@@ -2,6 +2,36 @@
 [![Maven Central](https://maven-badges.herokuapp.com/maven-central/io.skuber/skuber_2.12/badge.svg)](http://search.maven.org/#search|ga|1|g:%22io.skuber%22a:%22skuber_2.12%22)
 [![License](https://img.shields.io/badge/License-Apache%202.0-blue.svg)](https://github.com/doriordan/skuber/blob/master/LICENSE.txt)
 
+# 2.1 PREVIEW
+
+This is a preview of version 2.1 of Skuber. It contains the following changes:
+
+## New advanced controller API (ALPHA)
+
+** Currently incomplete **
+
+This new client API (currently in this [alpha package](client/src/main/scala/skuber/api/alpha/controller)) aims to provide similar functionality for Scala Kubernetes applications as that provided by the `client-go`  library under its `tools/cache` folder to Go Kubernetes applications, as described [here](https://github.com/kubernetes/sample-controller/blob/master/docs/controller-client-go.md)
+
+That `client-go` folder supports a set of components that reflects updates to resources on the cluster into a local (client-side) cache, and informs controllers of changes to those resources. The new Skuber package provides similar funcionality, albeit leveraging Akka streams to simplify and enhance the reactive handling of the flow of updates from the cluster through the cache and informers.  
+
+The primary use case is to make it easier to write custom controllers (as used for example by Kubernetes operators) that are reliably updated with changes on watched set of resources as well as benefitting from local caching of those resources.
+  
+## Refactoring of `skuber.api` package
+
+This refactoring should not require any changes to client applications but is a significant internal rewrite. The client API has now been factored out into a `KubernetesClient` trait, implemented by `KubernetesClientImpl` class (which is largely the same code as the previous `RequestContext` class, which is now jyst a tye alias to the new client implementation).
+
+Additionally other code in the package,, such the implementation of the pod `exec` method, has been factored out into separate packages/objects.
+
+Overall these changes greatly reduce the size pf the previously bloated `skuber.api.client` package object, and makes the code easier to understand and maintain.  
+
+## ListOptions
+ 
+Full support of [ListOptions](https://godoc.org/k8s.io/apimachinery/pkg/apis/meta/v1#ListOptions), which is a set of options that can be passed with list or watch commands (previously Skuber indirectly supported only some of those options).
+ 
+New `watchWithOptions` and `listWithOptions` client API methods enable any application-specified ListOptions to be passed with `watch` and `list` calls.
+
+Some of the existing `list` and `watch` calls have been internally refactored to use ListOptions in their implementation.
+ 
 # Skuber
 
 Skuber is a Scala client library for [Kubernetes](http://kubernetes.io). It provides a fully featured, high-level and strongly typed Scala API for managing Kubernetes cluster resources (such as Pods, Services, Deployments, ReplicaSets, Ingresses  etc.) via the Kubernetes REST API server.
