+++ client/src/main/scala/skuber/api/client/Cluster.scala
@@ -1,0 +1,13 @@
+package skuber.api.client
+
+/**
+  * @author David O'Riordan
+  *
+  * Defines the details needed to communicate with the API server for a Kubernetes cluster
+  */
+case class Cluster(
+  apiVersion: String = "v1",
+  server: String = defaultApiServerURL,
+  insecureSkipTLSVerify: Boolean = false,
+  certificateAuthority: Option[PathOrData] = None
+)
