+++ client/src/main/scala/skuber/api/client/Context.scala
@@ -1,0 +1,13 @@
+package skuber.api.client
+
+import skuber.Namespace
+
+/**
+  * @author David O'Riordan
+  * Define the Kubernetes API context for requests
+  */
+case class Context(
+  cluster: Cluster = Cluster(),
+  authInfo: AuthInfo = NoAuth,
+  namespace: Namespace = Namespace.default
+)
