+++ client/src/main/scala/skuber/api/watch/Watch.scala
@@ -1,12 +1,14 @@
-package skuber.api
+package skuber.api.watch
 
-import scala.concurrent.{ExecutionContext, Future}
-import scala.language.postfixOps
 import akka.http.scaladsl.model.{HttpMethods, _}
 import akka.stream.scaladsl.Source
 import play.api.libs.json.Format
 import skuber.api.client._
-import skuber.{ObjectResource, ResourceDefinition}
+import skuber.api.client.impl.KubernetesClientImpl
+import skuber.{ListOptions, ObjectResource, ResourceDefinition}
+
+import scala.concurrent.{ExecutionContext, Future}
+import scala.language.postfixOps
 
 /**
  * @author David O'Riordan
@@ -14,7 +16,7 @@ import skuber.{ObjectResource, ResourceDefinition}
  * Based on Akka streaming
  */
 object Watch {
-  
+
   /**
     * Get a source of events on a specific Kubernetes resource
     * @param context the applicable request context
@@ -26,13 +28,18 @@ object Watch {
     * @tparam O
     * @return
     */
-  def events[O <: ObjectResource](context: RequestContext, name: String, sinceResourceVersion: Option[String] = None, bufSize: Int)(
+  def events[O <: ObjectResource](context: KubernetesClientImpl, name: String, sinceResourceVersion: Option[String] = None, bufSize: Int)(
     implicit format: Format[O], rd: ResourceDefinition[O], lc: LoggingContext) : Future[Source[WatchEvent[O], _]] =
   {
     context.logInfo(context.logConfig.logRequestBasic, s"creating watch on resource $name of kind ${rd.spec.names.kind}")
 
-    val maybeResourceVersionQuery = sinceResourceVersion map { version => Uri.Query("resourceVersion" -> version) }
-    val request = context.buildRequest(HttpMethods.GET, rd, Some(name), query = maybeResourceVersionQuery, watch = true)
+    val nameFieldSelector=Some(s"metadata.name=$name")
+    val watchOptions=ListOptions(
+      resourceVersion = sinceResourceVersion,
+      watch = Some(true),
+      fieldSelector = nameFieldSelector
+    )
+    val request = context.buildRequest(HttpMethods.GET, rd, None, query = Some(Uri.Query(watchOptions.asMap)))
     val responseFut = context.invokeWatch(request)
     toFutureWatchEventSource(context, responseFut, bufSize)
   }
@@ -47,13 +54,13 @@ object Watch {
     * @tparam O
     * @return a Future which will eventually return a Source of events
     */
-  def eventsOnKind[O <: ObjectResource](context: RequestContext, sinceResourceVersion: Option[String] = None, bufSize: Int)(
+  def eventsOnKind[O <: ObjectResource](context: KubernetesClientImpl, sinceResourceVersion: Option[String] = None, bufSize: Int)(
     implicit format: Format[O], rd: ResourceDefinition[O], lc: LoggingContext) : Future[Source[WatchEvent[O], _]] =
   {
     context.logInfo(context.logConfig.logRequestBasic, s"creating skuber watch on kind ${rd.spec.names.kind}")
 
-    val maybeResourceVersionQuery = sinceResourceVersion map { v => Uri.Query("resourceVersion" -> v) }
-    val request = context.buildRequest(HttpMethods.GET, rd, None, query = maybeResourceVersionQuery, watch = true)
+    val watchOptions=ListOptions(resourceVersion = sinceResourceVersion, watch = Some(true))
+    val request = context.buildRequest(HttpMethods.GET, rd, None, query = Some(Uri.Query(watchOptions.asMap)))
     val responseFut = context.invokeWatch(request)
     toFutureWatchEventSource(context, responseFut, bufSize)
   }
@@ -67,7 +74,7 @@ object Watch {
     * @tparam O the Kubernetes kind of each events object
     * @return a Future which will eventually return a Source of events
     */
-  private def toFutureWatchEventSource[O <: ObjectResource](context: RequestContext, eventStreamResponseFut: Future[HttpResponse], bufSize: Int)(
+  private def toFutureWatchEventSource[O <: ObjectResource](context: KubernetesClientImpl, eventStreamResponseFut: Future[HttpResponse], bufSize: Int)(
     implicit format: Format[O],lc: LoggingContext): Future[Source[WatchEvent[O], _]] =
   {
     implicit val ec: ExecutionContext = context.actorSystem.dispatcher
